refactor: page.tsx에 있던 캐러셀 코드 별도의 컴포넌트로 분리
